not complete
i almost complete layout task, but i think it's perfect if i make it 'responsive'! but it's impossible now because i should use 'js language' to make that. sure i can make it using 'media query' in css, it's so inefficient work.....
And i didn't do turn background image when a tags are in hover state. because it's so hard.... must be modified!!!!!!!!